moved get_all_catalogs
